requirements.txt
torch>=2.3.0
PyYAML>=6.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
transformers>=4.51.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.12.0
bitsandbytes>=0.43.0
trl>=0.9.0
evaluate>=0.4.2
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=5.27.0
scikit-learn>=1.5.0
